backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pillow==10.3.0

# AI OCR
paddlepaddle==3.1.0
paddleocr==3.1.0
openai==1.99.1
deep-translator==1.11.4
